IMPASSE POINTt - Major stoppage point. Due to Generics. See deets
This is code stopping point that is at an impasse because due to the odd nature of the book using generic implementations that I was unsure how to re-implmenet in C#. We had to create a separate Interpreter that had a alot dupe code and re-use. However most of the fn signatures are geared toward the non-genericized interfaces of the IExprVisitor and ExprBase calsses. I would either need to continue with repeating myself- creating more diverging code ohere or go back and unify the interfaces. I dont know but it's creating a time based impasse. May call it quits here because I sort of understand many concepts of the interpreter at this point