Test with XOR operator for EQUAL_EQUAL comparison
I want to see if the XOR of each has and overall false thusly why the comparions has a negation. IF so then we should attemptto ue the original values to return that they are not fo the SAME TYPE to be tested.

Again I believe this is due ot CodeCrafots BS wher they made us in the earlier part of the courseforce numbers to have at least one .0 precision point that we need this ugly hack