Try implement hacky fix with a condition check for Doubles being interpreted as String
THis is again due to the differences in Java and C#  as well as the poor explanation of the book using Generics and me not implementing. Im paying the toll for that
In this method I attempt to reason about that I can possibly extend the isbothnums check to a newcheckto see  if both operands are false. So created a isbothfalse check. Theory is if theyre both false and werent parseable as Doubles then that indicates that theyre string types that should be concatenated. Though, I still dont know how close CodeCraft is following the Lox specification and any of theire surprise tests that might fail. So will commit and hopefully no further adjustments needto be made (except for possibly a change in the message in the thrown ParserErrorException