Remove the automatic true return use isEqual as expected. Additonally check against the base types.
I belive this function stopped working as intended because now we're explicitly parsing in the Tokenizer it breaks the general tests. Instead we pass directly to the previously never used isEqual function which for now seems to be working to return false when it compares a Double to a String.  This still checks "res" return value from the TrueForAll test because C# still doesnt consider it ToString reprenstations different