FINAL NON-WORKING STATE PUSH
Abandoning with this final trash upload that wont pass any tests due to me giving up with the convoluted hobbed along complexity of converting from Java to C# and having to split.  A bunch of NotImplemented are her just to get it to compile and watch it finally fail as I quit this. May transfer over to interpreterbook.co now as that uses simple GoLang and no strange implementations and restricing design decision. It was good while it lasted tho!